third_party/blink/web_tests/external/wpt/service-workers/service-worker/next-hop-protocol.https-expected.txt

This is a testharness.js-based test.
[FAIL] nextHopProtocol reports H2 correctly when routed via a service worker.
  assert_equals: nextHopProtocol is set on fallback expected "h2" but got "http/1.1"
Harness: the test ran to completion.
